requirements.txt
crewai[tools]
google-genai
playwright
pyperclip
pillow
piexif
sajupy
python-dotenv
requests
pyyaml
